src/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
cffi==1.15.1
Django==4.1
django-admin-blame==1.1.1
django-admin-interface==0.20.0
django-admin-log==0.2
django-ajax-selects==2.2.0
django-appconf==1.0.5
django-authtools==2.0.0
django-autocomplete-light==3.5.1
django-braces==1.15.0
django-captcha==1.0.0
django-classy-tags==3.0.1
django-colorfield==0.7.2
django-controlcenter==0.3.1
django-cookie-law==2.2.0
django-crispy-forms==1.14.0
django-crum==0.7.9
django-dbbackup==3.2.0
django-dbbackup-ui==0.2.1
django-debug-permissions==1.0.0
django-debug-toolbar==3.6.0
django-decouple==2.1
django-environ==0.9.0
django-extensions==3.2.0
django-filebrowser-no-grappelli==4.0.1
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-geojson==3.2.1
django-imagekit==4.1.0
django-italian-utils==0.3.3
django-leaflet==0.28.3
django-model-utils==4.2.0
django-ordered-model==3.6
django-phonenumber-field==6.4.0
django-phonenumbers==1.0.1
django-pkgconf==0.4.0
django-queryable-properties==1.8.3
django-recaptcha==3.0.0
django-related-admin==1.0.1
django-rest-framework==0.1.0
django-tables2==2.4.1
django-tinymce==3.5.0
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-datatables==0.7.0
djangorestframework-gis==1.0
easy-thumbnails==2.8.3
numpy==1.23.2
pandas==1.4.3
phonenumbers==8.12.54
pilkit==2.0
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
pytest-runner==6.0.0
python-dateutil==2.8.2
pytz==2022.2.1
sentry-sdk==1.9.5
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
unicodecsv==0.14.1
urllib3==1.26.12
